+++ src/sqlalchemy_declarative_extensions/role/__init__.py
@@ -1,5 +1,6 @@
-from sqlalchemy_declarative_extensions.role.generic import Role
+from sqlalchemy_declarative_extensions.role.generic import Env, Role
 
 __all__ = [
+    "Env",
     "Role",
 ]